% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_filt_data_combined.R
\name{generate_filt_data_combined}
\alias{generate_filt_data_combined}
\title{Generate Filt Data Combined}
\usage{
generate_filt_data_combined(filt_data_laser, filt_data_loupe)
}
\arguments{
\item{filt_data_laser}{}
}
\value{
A list
}
\description{
Generate Filt Data Combined
}
\examples{
filt_data_laser <- list(structure(list(Wavelength = NA, Lens = "Pi1", Summary = "High VLT, lightweight and relatively inexpensive",
OD = "800-830nm (OD 5), 9000-11000nm (OD 7)", VLT = "70\%",
Shape = "curved", Material = "Polycarbonate", Website = "https://innovativeoptics.com/pi1-laser-glasses-frames/",
Image = "https://innovativeoptics.com/wp-content/uploads/2018/04/pi1.jpg",
Graph = "graphs/Pi1_plot.png", `Price(from)` = "$175"), class = c("tbl_df",
                                                                 "tbl", "data.frame"), row.names = c(NA, -1L)))
filt_data_loupe <- structure(list(`Selected Device` = "Clarion (800-820nm OD 5)",
`Selected Loupes` = "Andau Blues LG (One size) ", `Compatible Loupe Insert` = "IVL.R"), row.names = c(NA,
                                                                                                     -1L), class = c("tbl_df", "tbl", "data.frame"))
generate_filt_data_combined(filt_data_laser, filt_data_loupe)
}
