% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lens_link.R
\name{generate_lens_link}
\alias{generate_lens_link}
\title{Generate Lens Link}
\usage{
generate_lens_link(dental_data, loupe_insert)
}
\arguments{
\item{dental_data}{}
}
\value{
A dataframe
}
\description{
Generate Lens Link
}
\examples{
dental_data <- tibble::tibble( `Laser Mfg` = "AMD", `Laser Model` = "Clarion",
Wavelengths = "800-820nm OD 5",`Laser Type` = "Diode Laser", `Eyewear Lens Compatible` = "Pi1",
Website = "https://innovativeoptics.com/pi1-laser-glasses-frames/")
loupe_insert <- tibble::tibble(Mfg = "Lumadent", `Lumadent Frame` = "Argon", Style = "One size",
`Innovative Optics Insert` = "IVL")
generate_lens_link(dental_data, loupe_insert)
}
